feat(area): pr fixes
