+++ antarest/study/business/model/area_model.py
@@ -35,7 +35,6 @@ class Area(AntaresBaseModel):
 class AreaCreation(AntaresBaseModel):
     name: str
     type: AreaType
-    set: Optional[List[str]] = None
 
 
 class UpdateAreaUi(AntaresBaseModel, extra="forbid", populate_by_name=True):
