+++ antarest/study/web/study_data_blueprint.py
@@ -13,7 +13,7 @@
 import enum
 import logging
 from http import HTTPStatus
-from typing import Any, Dict, List, Mapping, Optional, Sequence, cast
+from typing import Any, Dict, List, Mapping, Optional, Sequence
 
 import typing_extensions as te
 from fastapi import APIRouter, Body, Query
@@ -154,12 +154,12 @@ def create_study_data_routes(study_service: StudyService, config: Config) -> API
         "/studies/{uuid}/areas",
         tags=[APITag.study_data],
         summary="Get all areas basic info",
-        response_model=List[Area] | Dict[str, Any],
     )
     def get_areas(uuid: str, type: AreaType = Query(None), ui: bool = False) -> List[Area] | Dict[str, Any]:
         logger.info(f"Fetching area list (type={type}) for study {uuid}")
-        areas_list = study_service.get_all_areas(uuid, type, ui)
-        return areas_list
+        if ui:
+            return study_service.get_all_areas_ui_info(uuid)
+        return study_service.get_all_areas(uuid, type)
 
     @bp.get("/studies/{uuid}/links", tags=[APITag.study_data], summary="Get all links")
     def get_links(uuid: str) -> List[Link]:
@@ -171,9 +171,8 @@ def create_study_data_routes(study_service: StudyService, config: Config) -> API
         "/studies/{uuid}/areas",
         tags=[APITag.study_data],
         summary="Create a new area",
-        response_model=Area,
     )
-    def create_area(uuid: str, area_creation_info: AreaCreation) -> Any:
+    def create_area(uuid: str, area_creation_info: AreaCreation) -> Area:
         logger.info(f"Creating new area for study {uuid}")
         return study_service.create_area(uuid, area_creation_info)
 
@@ -1101,10 +1100,7 @@ def create_study_data_routes(study_service: StudyService, config: Config) -> API
         - the values are the allocation factors.
         """
         study = study_service.check_study_access(uuid, StudyPermissionType.READ)
-        all_areas = cast(
-            List[Area],
-            study_service.get_all_areas(uuid, area_type=AreaType.AREA, ui=False),
-        )
+        all_areas: List[Area] = study_service.get_all_areas(uuid, area_type=AreaType.AREA)
         study_interface = study_service.get_study_interface(study)
         return study_service.allocation_manager.get_allocation_matrix(study_interface, all_areas)
 
@@ -1125,10 +1121,7 @@ def create_study_data_routes(study_service: StudyService, config: Config) -> API
         Returns the allocation form fields.
         """
         study = study_service.check_study_access(uuid, StudyPermissionType.READ)
-        all_areas = cast(
-            List[Area],
-            study_service.get_all_areas(uuid, area_type=AreaType.AREA, ui=False),
-        )
+        all_areas: List[Area] = study_service.get_all_areas(uuid, area_type=AreaType.AREA)
         study_interface = study_service.get_study_interface(study)
         return study_service.allocation_manager.get_allocation_form_fields(all_areas, study_interface, area_id)
 
@@ -1162,10 +1155,7 @@ def create_study_data_routes(study_service: StudyService, config: Config) -> API
         Returns the updated allocation form fields.
         """
         study = study_service.check_study_access(uuid, StudyPermissionType.WRITE)
-        all_areas = cast(
-            List[Area],
-            study_service.get_all_areas(uuid, area_type=AreaType.AREA, ui=False),
-        )
+        all_areas: List[Area] = study_service.get_all_areas(uuid, area_type=AreaType.AREA)
         study_interface = study_service.get_study_interface(study)
         return study_service.allocation_manager.set_allocation_form_fields(all_areas, study_interface, area_id, data)
 
@@ -1211,10 +1201,7 @@ def create_study_data_routes(study_service: StudyService, config: Config) -> API
         - `data`: a 2D-array matrix of correlation coefficients with values in the range of -1 to 1.
         """
         study = study_service.check_study_access(uuid, StudyPermissionType.READ)
-        all_areas = cast(
-            List[Area],
-            study_service.get_all_areas(uuid, area_type=AreaType.AREA, ui=False),
-        )
+        all_areas: List[Area] = study_service.get_all_areas(uuid, area_type=AreaType.AREA)
         study_interface = study_service.get_study_interface(study)
         return study_service.correlation_manager.get_correlation_matrix(
             all_areas,
@@ -1257,10 +1244,7 @@ def create_study_data_routes(study_service: StudyService, config: Config) -> API
         Returns the hydraulic/load/solar/wind correlation matrix updated
         """
         study = study_service.check_study_access(uuid, StudyPermissionType.WRITE)
-        all_areas = cast(
-            List[Area],
-            study_service.get_all_areas(uuid, area_type=AreaType.AREA, ui=False),
-        )
+        all_areas: List[Area] = study_service.get_all_areas(uuid, area_type=AreaType.AREA)
         study_interface = study_service.get_study_interface(study)
         return study_service.correlation_manager.set_correlation_matrix(all_areas, study_interface, matrix)
 
@@ -1281,10 +1265,7 @@ def create_study_data_routes(study_service: StudyService, config: Config) -> API
         Returns the correlation form fields in percentage.
         """
         study = study_service.check_study_access(uuid, StudyPermissionType.READ)
-        all_areas = cast(
-            List[Area],
-            study_service.get_all_areas(uuid, area_type=AreaType.AREA, ui=False),
-        )
+        all_areas: List[Area] = study_service.get_all_areas(uuid, area_type=AreaType.AREA)
         study_interface = study_service.get_study_interface(study)
         return study_service.correlation_manager.get_correlation_form_fields(all_areas, study_interface, area_id)
 
@@ -1318,10 +1299,7 @@ def create_study_data_routes(study_service: StudyService, config: Config) -> API
         Returns the correlation form fields in percentage.
         """
         study = study_service.check_study_access(uuid, StudyPermissionType.WRITE)
-        all_areas = cast(
-            List[Area],
-            study_service.get_all_areas(uuid, area_type=AreaType.AREA, ui=False),
-        )
+        all_areas: List[Area] = study_service.get_all_areas(uuid, area_type=AreaType.AREA)
         study_interface = study_service.get_study_interface(study)
         return study_service.correlation_manager.set_correlation_form_fields(all_areas, study_interface, area_id, data)
